requirements.txt
fpdf
python-dotenv
pyttsx3
google-generativeai
youtube-transcript-api
python-docx
